+++ src/main/kotlin/localization/Strings.kt
@@ -82,5 +82,8 @@ data class Strings(
     val waitingForChanges: String,
     val iAmDone: String,
     val mergingSteamManifest: (name: String) -> String,
-    val reLinkingInstances: String
+    val reLinkingInstances: String,
+    val save: String,
+    val editInstanceTitle: (name: String) -> String,
+    val editBaseInstance: String
 )